refactor: extract reusable LinkList component from settings page
- Create generic LinkList component for navigation lists
- Move link list rendering logic to components directory
- Update settings page to use new LinkList component
- Enable reuse of link list pattern across other pages
